# MNT Optimize safe_indexing for slices