This PR uses `islice` instead of creating the whole list and slicing it. Here is a quick benchmark:

```python
from itertools import islice

%%timeit
_ = list(range(n_columns))[slice(start, end)]
76.7 µs ± 409 ns per loop

%%timeit
_ = list(islice(range(n_columns), start, stop))
3.88 µs ± 40.6 ns per loop
```